ml/requirements.txt
flask
flask-cors
tensorflow>=2.12
opencv-python-headless
pillow
numpy
scikit-image
